# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Conjur API
  description: OpenAPI specification for Conjur - compatible with CyberArk Conjur
  version: "5.0.0"
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Basic authentication with username and API key
    bearerAuth:
      type: http
      scheme: bearer
      description: Conjur access token
  schemas:
...
    PolicyLoadResponse:
      type: object
      properties:
        created_roles:
          type: object
          additionalProperties:
            type: object
            properties:
              id:
                type: string
              api_key:
                type: string
        version:
          type: integer
        dry_run:
          type: boolean
    PolicyVersion:
      type: object
      properties:
        version:
          type: integer
        created_at:
          type: string
          format: date-time
        policy_sha256:
          type: string
        finished_at:
          type: string
          format: date-time
        client_ip:
          type: string
        role_id:
          type: string
    Resource:
      type: object
      properties:
        id:
          type: string
        owner:
          type: string
        policy:
          type: string
        permissions:
          type: array
          items:
            type: object
        annotations:
          type: object
          additionalProperties:
            type: string
    Role:
      type: object
      properties:
        id:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/MembershipGrant"
    MembershipGrant:
      type: object
      properties:
        role:
          type: string
        member:
          type: string
        admin_option:
          type: boolean
        ownership:
          type: boolean
        policy:
          type: string
  parameters:
    Account:
      name: account
      in: path
      required: true
      schema:
        type: string
      description: Organization account name
...
    Kind:
      name: kind
      in: path
      required: true
      schema:
        type: string
      description: Resource kind
    Identifier:
      name: identifier
      in: path
      required: true
      schema:
        type: string
      description: URL-encoded resource identifier
...
paths:
...
  /secrets/{account}/{kind}/{identifier}:
    get:
      operationId: getSecret
      summary: Fetch a secret value
      tags: [secrets]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Account"
        - $ref: "#/components/parameters/Kind"
        - $ref: "#/components/parameters/Identifier"
        - name: version
          in: query
          schema:
            type: integer
          description: Secret version (default latest)
      responses:
        "200":
          description: Secret value
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "403":
          description: Forbidden
        "404":
          description: Secret not found

    post:
      operationId: createSecret
      summary: Create or update a secret value, or expire it
      tags: [secrets]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Account"
        - $ref: "#/components/parameters/Kind"
        - $ref: "#/components/parameters/Identifier"
        - name: expirations
          in: query
          schema:
            type: string
          description: If present, expire the secret instead of setting value
      requestBody:
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Secret created/expired
        "403":
          description: Forbidden
        "422":
          description: Invalid kind for expiration
...
  /policies/{account}/policy/{identifier}:
    get:
      operationId: getPolicy
      summary: Get policy versions
      tags: [policies]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Account"
        - $ref: "#/components/parameters/Identifier"
        - name: version
          in: query
          schema:
            type: integer
          description: Specific version to retrieve
      responses:
        "200":
          description: Policy versions or policy text
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PolicyVersion"
            application/x-yaml:
              schema:
                type: string
        "403":
          description: Forbidden

    post:
      operationId: loadPolicy
      summary: Append to existing policy
      tags: [policies]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Account"
        - $ref: "#/components/parameters/Identifier"
        - name: dry_run
          in: query
          schema:
            type: boolean
          description: Validate without applying
      requestBody:
        required: true
        content:
          application/x-yaml:
            schema:
              type: string
          text/plain:
            schema:
              type: string
      responses:
        "201":
          description: Policy loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyLoadResponse"
        "200":
          description: Dry run successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyLoadResponse"
        "422":
          description: Invalid policy

    put:
      operationId: replacePolicy
      summary: Replace existing policy
      tags: [policies]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Account"
        - $ref: "#/components/parameters/Identifier"
        - name: dry_run
          in: query
          schema:
            type: boolean
      requestBody:
        required: true
        content:
          application/x-yaml:
            schema:
              type: string
          text/plain:
            schema:
              type: string
      responses:
        "201":
          description: Policy replaced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyLoadResponse"
        "200":
          description: Dry run successful
        "422":
          description: Invalid policy

    patch:
      operationId: updatePolicy
      summary: Update existing policy
      tags: [policies]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Account"
        - $ref: "#/components/parameters/Identifier"
        - name: dry_run
          in: query
          schema:
            type: boolean
      requestBody:
        required: true
        content:
          application/x-yaml:
            schema:
              type: string
          text/plain:
            schema:
              type: string
      responses:
        "201":
          description: Policy updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyLoadResponse"
        "200":
          description: Dry run successful
        "422":
          description: Invalid policy

  # ========== RESOURCES ==========
...
  /resources/{account}/{kind}/{identifier}:
    get:
      operationId: getResource
      summary: Get a specific resource
      tags: [resources]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Account"
        - $ref: "#/components/parameters/Kind"
        - $ref: "#/components/parameters/Identifier"
        - name: check
          in: query
          schema:
            type: boolean
          description: If true, check permission instead of fetching resource
        - name: privilege
          in: query
          schema:
            type: string
          description: Required with check=true
        - name: role
          in: query
          schema:
            type: string
          description: Role to check permission for
        - name: permitted_roles
          in: query
          schema:
            type: boolean
          description: If present, return roles with the specified privilege
      responses:
        "200":
          description: Resource details or permitted roles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "204":
          description: Permission check passed
        "403":
          description: Forbidden
        "404":
          description: Resource not found

  # ========== ROLES ==========
  /roles/{account}/{kind}/{identifier}:
    get:
      operationId: getRole
      summary: Get role details
      tags: [roles]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Account"
        - $ref: "#/components/parameters/Kind"
        - $ref: "#/components/parameters/Identifier"
        - name: members
          in: query
          schema:
            type: boolean
          description: If present, return role members
        - name: memberships
          in: query
          schema:
            type: boolean
          description: If present, return role memberships
        - name: all
          in: query
          schema:
            type: boolean
          description: If present with memberships, return all transitive memberships
        - name: count
          in: query
          schema:
            type: boolean
          description: Return count instead of list
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Role details, members, or memberships
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Role"
                  - type: array
                    items:
                      $ref: "#/components/schemas/MembershipGrant"
                  - type: array
                    items:
                      type: string
                  - type: object
                    properties:
                      count:
                        type: integer
        "404":
          description: Role not found

    post:
      operationId: addRoleMember
      summary: Add member to role
      tags: [roles]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Account"
        - $ref: "#/components/parameters/Kind"
        - $ref: "#/components/parameters/Identifier"
        - name: members
          in: query
          required: true
          schema:
            type: string
          description: Required flag to indicate member operation
        - name: member
          in: query
          required: true
          schema:
            type: string
          description: Member role ID to add
      responses:
        "204":
          description: Member added
        "403":
          description: Forbidden
        "404":
          description: Role or member not found

    delete:
      operationId: deleteRoleMember
      summary: Remove member from role
      tags: [roles]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Account"
        - $ref: "#/components/parameters/Kind"
        - $ref: "#/components/parameters/Identifier"
        - name: members
          in: query
          required: true
          schema:
            type: string
          description: Required flag to indicate member operation
        - name: member
          in: query
          required: true
          schema:
            type: string
          description: Member role ID to remove
      responses:
        "204":
          description: Member removed
        "403":
          description: Forbidden
        "404":
          description: Membership not found